requirements.txt
google-cloud-speech==2.21.0
google-cloud-translate==3.11.1
pyaudio==0.2.13